Домашнее задание 1 урока.

+++ homework1/src/main/java/product/Product.java
@@ -8,5 +8,5 @@ import lombok.Getter;
 public class Product {
     private long id;
     private String title;
-    private double cost;
+    private int cost;
 }
